podcast/requirements.txt
elevenlabs==1.12.1
openai==1.54.3
python-dotenv==1.0.1
